refactor(build): make src fully TypeScript (index + types)
Convert the package entry to src/index.ts, rename src/types.d.ts to
src/types.ts (imported via .js specifiers), and point the root export
at the TypeScript source. No remaining .js or .d.ts under src.
