+++ src/Func/curry.ts
@@ -1,4 +1,4 @@
-import type { AnyFn, Curried } from '../types.d.ts'
+import type { AnyFn, Curried } from '../types.js'
 
 /**
  * Returns a curried version of the supplied function.
