+++ src/List/zip.ts
@@ -1,4 +1,4 @@
-import type { Pair } from '../types.d.ts'
+import type { Pair } from '../types.js'
 import curry from '../Func/curry.js'
 
 /**
