+++ src/Str/drop.ts
@@ -1,4 +1,4 @@
-import type { Curried } from '../types.d.ts'
+import type { Curried } from '../types.js'
 import dropList from '../List/drop.js'
 import curry from '../Func/curry.js'
