+++ src/Str/dropWhile.ts
@@ -1,4 +1,4 @@
-import type { Curried } from '../types.d.ts'
+import type { Curried } from '../types.js'
 import dropWhileList from '../List/dropWhile.js'
 import curry from '../Func/curry.js'
