+++ src/Str/splitAt.ts
@@ -1,4 +1,4 @@
-import type { Curried } from '../types.d.ts'
+import type { Curried } from '../types.js'
 import curry from '../Func/curry.js'
 import drop from './drop.js'
 import take from './take.js'
