--- src/index.js
@@ -1,19 +1,0 @@
-/**
- * A modular, tree-shaking friendly implementation of Haskell's Prelude library in modern JavaScript.
- *
- * This package exports the following modules:
- * - {@link Func}
- * - {@link General}
- * - {@link List}
- * - {@link Num}
- * - {@link Obj}
- * - {@link Str}
- *
- * @module
- */
-export * as Func from './Func/index.js'
-export * as General from './General/index.js'
-export * as List from './List/index.js'
-export * as Num from './Num/index.js'
-export * as Obj from './Obj/index.js'
-export * as Str from './Str/index.js'
