+++ tsconfig.types.json
@@ -14,5 +14,5 @@
     "target": "ES2022",
     "types": ["bun"]
   },
-  "include": ["src/**/*.ts", "src/**/*.d.ts", "test/**/*.test.ts"]
+  "include": ["src/**/*.ts", "test/**/*.test.ts"]
 }
